+++ app/Plugin/Settings.php
@@ -13,7 +13,9 @@ defined( 'ABSPATH' ) || exit;
 use DownloadListWithIcons\Dependencies\easySettingsForWordPress\Fields\Button;
 use DownloadListWithIcons\Dependencies\easySettingsForWordPress\Fields\Checkbox;
 use DownloadListWithIcons\Dependencies\easySettingsForWordPress\Fields\Select;
+use DownloadListWithIcons\Dependencies\easySettingsForWordPress\Fields\SelectPostTypeObject;
 use DownloadListWithIcons\Dependencies\easySettingsForWordPress\Fields\Text;
+use DownloadListWithIcons\Dependencies\easyTransientsForWordPress\Transients;
 use DownloadListWithIcons\Iconsets\Iconsets;
 use WP_Query;
 
@@ -121,7 +123,6 @@ class Settings {
 		$tasks_tab = $settings_page->add_tab( 'downloadlist_tasks', 20 );
 		$tasks_tab->set_name( 'downloadlist_tasks' );
 		$tasks_tab->set_title( __( 'Tasks', 'download-list-block-with-icons' ) );
-		$tasks_tab->set_hide_save( true );
 
 		// the helper tab.
 		$helper_tab = $settings_page->add_tab( 'downloadlist_helper', 70 );
@@ -134,7 +135,7 @@ class Settings {
 		 */
 		// the info section.
 		$general_tab_info = $general_tab->add_section( 'info', 10 );
-		$general_tab_info->set_title( __( 'Information', 'download-list-block-with-icons' ) );
+		$general_tab_info->set_title( _x( 'How to use', 'Settings', 'download-list-block-with-icons' ) );
 		$general_tab_info->set_callback( array( $this, 'show_info' ) );
 		$general_tab_info->set_setting( $settings_obj );
 
@@ -394,16 +395,50 @@ class Settings {
 		);
 
 		// add setting.
-		$download_button_setting = $settings_obj->add_setting( 'downloadlist_inherit_settings' );
-		$download_button_setting->set_section( $tasks_tab_tasks );
-		$download_button_setting->prevent_export( true );
+		$setting = $settings_obj->add_setting( 'downloadlist_inherit_settings' );
+		$setting->set_section( $tasks_tab_tasks );
+		$setting->prevent_export( true );
 		$field = new Button();
 		$field->set_title( __( 'Inherit settings', 'download-list-block-with-icons' ) );
 		$field->set_button_url( '#' );
 		$field->set_button_title( __( 'Inherit settings to all download list blocks', 'download-list-block-with-icons' ) );
 		$field->add_data( 'dialog', Helper::get_json( $dialog ) );
 		$field->add_class( 'easy-dialog-for-wordpress' );
-		$download_button_setting->set_field( $field );
+		$setting->set_field( $field );
+
+		// add setting.
+		$setting = $settings_obj->add_setting( 'downloadlist_add_block_to_page' );
+		$setting->set_section( $tasks_tab_tasks );
+		$setting->prevent_export( true );
+		$setting->set_save_callback( array( $this, 'add_block_in_page' ) );
+		$field = new SelectPostTypeObject();
+		$field->set_title( __( 'Add block to page', 'download-list-block-with-icons' ) );
+		$field->set_button_title( __( 'Choose a page', 'download-list-block-with-icons' ) );
+		$field->set_popup_title( __( 'Choose and select a page', 'download-list-block-with-icons' ) );
+		$field->set_endpoint( '/wp-json/wp/v2/pages' );
+		$field->set_limit( 10 );
+		$field->set_chosen_title( __( 'Chosen page', 'download-list-block-with-icons' ) );
+		$field->set_label_title( __( 'Search for a page', 'download-list-block-with-icons' ) );
+		$field->set_placeholder( __( 'Enter the name of a page', 'download-list-block-with-icons' ) );
+		$field->set_cancel_button_title( __( 'Cancel', 'download-list-block-with-icons' ) );
+		$setting->set_field( $field );
+
+		// add setting.
+		$setting = $settings_obj->add_setting( 'downloadlist_add_block_to_post' );
+		$setting->set_section( $tasks_tab_tasks );
+		$setting->prevent_export( true );
+		$setting->set_save_callback( array( $this, 'add_block_in_post' ) );
+		$field = new SelectPostTypeObject();
+		$field->set_title( __( 'Add block to post', 'download-list-block-with-icons' ) );
+		$field->set_button_title( __( 'Choose a post', 'download-list-block-with-icons' ) );
+		$field->set_popup_title( __( 'Choose and select a post', 'download-list-block-with-icons' ) );
+		$field->set_endpoint( '/wp-json/wp/v2/posts' );
+		$field->set_limit( 10 );
+		$field->set_chosen_title( __( 'Chosen post', 'download-list-block-with-icons' ) );
+		$field->set_label_title( __( 'Search for a post', 'download-list-block-with-icons' ) );
+		$field->set_placeholder( __( 'Enter the name of a post', 'download-list-block-with-icons' ) );
+		$field->set_cancel_button_title( __( 'Cancel', 'download-list-block-with-icons' ) );
+		$setting->set_field( $field );
 
 		// initialize this settings object.
 		$settings_obj->init();
@@ -615,6 +650,104 @@ class Settings {
 	 * @return void
 	 */
 	public function show_info(): void {
-		echo '<p>' . esc_html__( 'Set the default settings for new download list blocks here. Under “Tasks,” you can also inherit these settings to all existing blocks.', 'download-list-block-with-icons' ) . '</p>';
+		/* translators: %1$s will be replaced by a URL. */
+		echo '<p>' . wp_kses_post( sprintf( __( 'Set the default settings for new download list blocks here. Under <a href="%1$s">Tasks</a>, you can also inherit these settings to all existing blocks.', 'download-list-block-with-icons' ), $this->get_url( 'downloadlist_tasks' ) ) ) . '</p>';
+	}
+
+	/**
+	 * Add block to given page.
+	 *
+	 * It will be added at the end of the content of the given page.
+	 *
+	 * @param string $values The value.
+	 * @return void
+	 */
+	public function add_block_in_page( string $values ): void {
+		// convert to int.
+		$post_id = absint( $values );
+
+		// bail if no page ID is given.
+		if ( 0 === $post_id ) {
+			return;
+		}
+
+		// add the block.
+		$this->add_block( $post_id );
+
+		// get the edit URL for this page.
+		$edit_url = get_edit_post_link( $post_id );
+		if ( ! is_string( $edit_url ) ) {
+			return;
+		}
+
+		// get its title.
+		$title = get_the_title( $post_id );
+
+		// add success message.
+		$transient_obj = Transients::get_instance()->add();
+		$transient_obj->set_name( 'downloadlist_block_added' );
+		$transient_obj->set_type( 'success' );
+		/* translators: %1$s will be replaced by a string. */
+		$transient_obj->set_message( '<strong>' . sprintf( __( 'The block has been added to %1$s.', 'download-list-block-with-icons' ), '<em>' . $title . '</em>' ) . '</strong> ' . sprintf( __( '<a href="%1$s">Edit the page</a> and add the files you want to list to the block.', 'download-list-block-with-icons' ), $edit_url ) );
+		$transient_obj->save();
+	}
+
+	/**
+	 * Add block to given post.
+	 *
+	 * It will be added at the end of the content of the given post.
+	 *
+	 * @param string $values The value.
+	 * @return void
+	 */
+	public function add_block_in_post( string $values ): void {
+		// convert to int.
+		$post_id = absint( $values );
+
+		// bail if no page ID is given.
+		if ( 0 === $post_id ) {
+			return;
+		}
+
+		// add the block.
+		$this->add_block( $post_id );
+
+		// get the edit URL for this page.
+		$edit_url = get_edit_post_link( $post_id );
+		if ( ! is_string( $edit_url ) ) {
+			return;
+		}
+
+		// get its title.
+		$title = get_the_title( $post_id );
+
+		// add success message.
+		$transient_obj = Transients::get_instance()->add();
+		$transient_obj->set_name( 'downloadlist_block_added' );
+		$transient_obj->set_type( 'success' );
+		/* translators: %1$s will be replaced by a string. */
+		$transient_obj->set_message( '<strong>' . sprintf( __( 'The block has been added to the post %1$s.', 'download-list-block-with-icons' ), '<em>' . $title . '</em>' ) . '</strong> ' . sprintf( __( '<a href="%1$s">Edit the post</a> and add the files you want to list to the block.', 'download-list-block-with-icons' ), $edit_url ) );
+		$transient_obj->save();
+	}
+
+	/**
+	 * Add block to any post-entry.
+	 *
+	 * @param int $post_id The post ID to use.
+	 * @return void
+	 */
+	private function add_block( int $post_id ): void {
+		// get the content.
+		$content = get_post_field( 'post_content', $post_id, 'raw' );
+
+		// add our own block at the end.
+		$content .= '<!-- wp:downloadlist/list /-->';
+
+		// save the content.
+		$query = array(
+			'ID'           => $post_id,
+			'post_content' => $content,
+		);
+		wp_update_post( $query );
 	}
 }